+++ web/app/Http/Routers/Admin/index.router.php
@@ -1,4 +1,4 @@
 <?php
 Route::group(["prefix"=>"/admin",'namespace' => 'Admin'], function(){
-    Route::get('/', "PassportController@Index");
+    Route::get('/', "IndexController@Index");
 });